marketing-strategy-agent/requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# AI & ML
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-openai>=0.0.5
langgraph>=0.0.26
openai>=1.12.0
google-generativeai>=0.3.2
tavily-python>=0.3.0

# Database & Vector Store
tidb-vector>=0.0.4
pymysql>=1.1.0
cryptography>=41.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Web & HTTP
httpx>=0.26.0
aiohttp>=3.9.0
requests>=2.31.0
jinja2>=3.1.2
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.9.0
tenacity>=8.2.0

# Logging & Monitoring
structlog>=23.2.0
python-json-logger>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development
black>=23.12.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0

# Report Generation
reportlab>=4.0.0

# Optional: Social Media APIs (uncomment as needed)
# tweepy==4.14.0
# facebook-business==18.0.2
# linkedin-api==2.0.1a3

# Optional: SEO Tools (uncomment as needed)
# google-api-python-client==2.110.0
# serpapi==0.1.5